feat: move book details to dedicated page route
