+++ src/components/index.ts
@@ -1,5 +1,4 @@
 export { default as BookCard } from "./BookCard";
 export { default as AddBookDialog } from "./AddBookDialog";
-export { default as BookDetailModal } from "./BookDetailModal";
 export { default as AppLayout } from "./AppLayout";
 export { default as SetPasswordDialog } from "./SetPasswordDialog";
