+++ src/pages/index.ts
@@ -2,3 +2,4 @@ export { default as Login } from "./Login";
 export { default as Dashboard } from "./Dashboard";
 export { default as Library } from "./Library";
 export { default as Analytics } from "./Analytics";
+export { default as BookDetails } from "./BookDetails";
